requirements.txt
python-dotenv>=0.19.2
requests>=2.24.0
numpy>=1.22.3
pandas>=1.4.1
scipy==1.8.0
Nasdaq-Data-Link==1.0.2
python-binance>=1.0.16
matplotlib==3.5.2
matplotlib-inline==0.1.3
backtrader==1.9.76.123
pandas-ta==0.3.14b0
wrapt==1.14.1
pytest==7.1.1
pytest-runner==6.0.0
ipykernel==6.15.1
setuptools>=60.5.0
wheel>=0.37.1
mkinit==0.3.4
tabulate==0.8.10
ipympl==0.9.2

# Include install_requires from setup.py in --editable mode (development)
# -e .
# Include install_requires from setup.py
# . 

# backtrader patch: https://stackoverflow.com/a/67740176
# -e git+https://github.com/mementum/backtrader.git@0fa63ef4a35dc53cc7320813f8b15480c8f85517#egg=backtrader
